Add temporary workaround for Windows pipe support of swift-system
The support for `pipe` on Windows was added by https://github.com/apple/swift-system/commit/cfd3f9da906cf62efafd162d524df6334cc85da6
but it is not included in the latest swift-system release 1.3.1.
This commit adds a temporary inlined implementation of `pipe` API for
Windows until we update to the next release of swift-system.