[docs] NFC: Remove redundant `--sccache` flag (#75390)
Following along with the Getting Started guide. I noticed that the guide specifically suggests adding the `--sccache` flag if needed, but it is included in the example invocation:

> If you installed and want to use Sccache, add --sccache to the invocation.

This means that the flag is optional and should be removed from the example invocation leaving at the user discretion to use.