[IRGen] Add a direct reference to Obj-C classes in their symbolic references.
In some cases, users may want to intentionally link certain Obj-C dependencies
without `-ObjC` or `-force_load`. When Swift references a generic Obj-C class,
it uses the mangled name type metadata accessors, and no direct reference to
the class symbol is emitted in the object code. (This differs from Clang, where
naming the class in source does emit such a reference.)

This change appends a pointer to that symbol after the null terminator of the
mangled symbolic reference. It will therefore not be considered part of the
name, but will keep the symbol alive for the linker.

Context: https://forums.swift.org/t/odd-behavior-involving-generic-obj-c-classes-and-type-metadata-accessors/73265