Deprecate versioned `__swift__` macro in imported (Obj)C(++) code.
Leaving behind an un-versioned `__swift__` macro.
Originally added in https://github.com/swiftlang/swift/pull/4510 (rdar://26921435) following a discussion in https://lists.swift.org/pipermail/swift-dev/Week-of-Mon-20160822/002754.html, the versioned macro was used to allow clients primarily to ready their existing ObjectiveC code for Swift 2 -> Swift 3 transition.

Now that neither one of those versions is a supported language mode and we have versioned APINotes support, and we do not currently know of clients that rely on the version carried by this macro, we would like to deprecate it, leaving behind the un-versioned `-D__swift__`.

The reason for the deprecation is that it is likely that in a build with multiple targets which have very similar configuration aside from Swift language version, they would not be able to share module dependencies due to the mismatch in the macro value.