Sema: Fix crash with property override and almost-valid superclass
It's possible that getSuperclassDecl() returns something but
getSuperclass() does not, for example if you reference a generic
superclass with missing or invalid generic arguments. We would
crash in that case when going down this particular code path.

However, the call to getSuperclass() and the entire function it
appeared in was actually unnecessary.

Fixes https://github.com/swiftlang/swift/issues/74651
Fixes rdar://problem/130394943