🍒[6.0][cxx-interop] Fix reverse interop crash when using raw modules
Explanation: A low-level clang API was used that relies on a cached
value being present. This fix is changing that API call to another that
calculates the value if the cache is not yet populated.
Scope: C++ reverse interop.
Risk: Low, replacing an API call with a different that has the same
semantics (other than calculating the value when the cache is empty).
Testing: Regression test added.
Issue: rdar://132746445
Reviewer: @egorzhdan
Original PR: #75714