[Concurrency] Suggest adding a method, when mutating actor property cross isolation
This comes as an idea from the [Why do not actor-isolated properties
support 'await'
setter](https://forums.swift.org/t/why-do-not-actor-isolated-properties-support-await-setter/73867/26) thread.

The note that
> Actor-isolated property 'colorFill' can not be mutated from the main actor

Leads people to assume that this is somehow "broken" while they could
add a metho to perform the mutation and move on.

A developer comments that:

> So I just assumed that this meant that it was fundamentally wrong to
> mutate this actor's properties from another actor, while all I actually
> needed was... a 3 line setter.