[cxx-interop][windows] support using a custom libc++ on Windows
This change builds upon the support for specifying which C++ standard
library to use, by adding support for using a custom libc++ on Windows.

Even though the clang Windows driver doesn't support libc++ explicitly,
Swift can still use it by allowing users to pass in the path to libc++
using the -cxx-isystem flag.

The CxxStdlib module is prohibited to use on Windows when libc++ is
used, as it has incompatible ABI.