NFC: generalize canAddExplicitConsume type queries
I'd like to be able to safely query whether we can add an explicit
consume around an Expr, even in the middle of applying a
ConstraintSystem solution.

In order to do that safely, I believe we need to query the
ConstraintSystem for the type of an Expr. So, this patch generalizes the
one 'TypeBase::getType' call to instead use the function argument to
query for types.